完整版

+++ app/src/main/java/com/example/user/guokun/bean/CouponBean.java
@@ -125,7 +125,7 @@ public class CouponBean {
             private String update_time;
             private int type;
             private String name;
-            private int money;
+            private double money;
             private String start_time;
             private String end_time;
             private int status;
@@ -170,11 +170,11 @@ public class CouponBean {
                 this.name = name;
             }
 
-            public int getMoney() {
+            public double getMoney() {
                 return money;
             }
 
-            public void setMoney(int money) {
+            public void setMoney(double money) {
                 this.money = money;
             }
